+++ react-components/src/components/CountryChooser.jsx
@@ -1,95 +1,38 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import PropTypes from 'prop-types'
 import Select from 'react-select'
-import ErrorList from '@src/components/ErrorList'
-import CountryData from '@src/components/CountryData'
-import Services from '../Services'
-import { slugify } from '../Helpers'
 import './stylesheets/CountryChooser.scss'
 
-class CountryChooser extends React.Component {
+export default class CountryChooser extends React.Component {
   constructor(props) {
     super(props)
 
-    const { selectedCountries } = this.props
-
-    const updatedSelectedCountryList = this.sanitizeSelectedCountries(selectedCountries)
-    const updatedCountryList = this.updatedCountryList(updatedSelectedCountryList)
+    const { countryList } = this.props
 
     this.state = {
       open: false,
-      loading: false,
-      selectedCountries: updatedSelectedCountryList,
       selectedCountry: null,
-      countryList: updatedCountryList,
-      errors: {},
+      countryList,
     }
 
     this.showCountrySelect = this.showCountrySelect.bind(this)
-    this.addCountry = this.addCountry.bind(this)
-    this.handleGetCountryDataSuccess = this.handleGetCountryDataSuccess.bind(this)
-    this.handleGetCountryDataError = this.handleGetCountryDataError.bind(this)
   }
 
-  removeCountry = (country) => {
-    const { selectedCountries } = this.state
-    const updatedSelectedCountries = selectedCountries.filter((item) => item !== country)
-    const updatedCountryList = this.updatedCountryList(updatedSelectedCountries)
-    this.setState({ selectedCountries: updatedSelectedCountries, countryList: updatedCountryList })
-    return false
+  componentDidUpdate(newProps) {
+    const { countryList } = this.state
+    if (newProps.countryList !== countryList) {
+      this.updateCountryList(newProps.countryList)
+      this.changeCountry(null)
+    }
   }
 
   changeCountry = (country) => {
     this.setState({ selectedCountry: country })
   }
 
-  sanitizeSelectedCountries = (array) => {
-    return array.filter((country) => country.export_duty !== undefined)
-  }
-
-  updatedCountryList = (selectedCountries) => {
-    const { countryList } = this.props
-    return countryList.filter(
-      (country) => selectedCountries.filter((selectedCountry) => selectedCountry.country === country.value).length === 0
-    )
-  }
-
-  addCountry() {
-    const { selectedCountry, selectedCountries } = this.state
-    const isExisting = selectedCountries.filter((country) => country.country === selectedCountry.value).length > 0
-
-    if (!isExisting) {
-      this.setState({
-        loading: true,
-        errors: {},
-      })
-
-      Services.getCountryData(selectedCountry.label)
-        .then(this.handleGetCountryDataSuccess)
-        .catch(this.handleGetCountryDataError)
-    }
-  }
-
-  handleGetCountryDataSuccess(data) {
-    // data should return only a single country
-    // currently it returns the whole array of selected countries
-    // TODO needs BE work
-    const updatedSelectedCountryList = this.sanitizeSelectedCountries(data.target_markets)
-    const updatedCountryList = this.updatedCountryList(updatedSelectedCountryList)
-
+  updateCountryList(list) {
     this.setState({
-      errors: {},
-      loading: false,
-      selectedCountries: updatedSelectedCountryList,
-      countryList: updatedCountryList,
-    })
-  }
-
-  handleGetCountryDataError(errors) {
-    this.setState({
-      errors: errors.message || errors,
-      loading: false,
+      countryList: list,
     })
   }
 
@@ -99,8 +42,9 @@ class CountryChooser extends React.Component {
   }
 
   render() {
-    const { selectedCountry, loading, open, selectedCountries, errors } = this.state
-    const { countryList } = this.state
+    const { selectedCountry, loading, open, countryList } = this.state
+    const { addCountry } = this.props
+
     let saveButton
     if (selectedCountry) {
       saveButton = (
@@ -108,7 +52,7 @@ class CountryChooser extends React.Component {
           type="button"
           className="country-chooser-save-button"
           id="country-chooser-save-button"
-          onClick={this.addCountry}
+          onClick={() => addCountry(selectedCountry)}
           disabled={loading}
         >
           Add
@@ -141,27 +85,8 @@ class CountryChooser extends React.Component {
       )
     }
 
-    let loadingMessage
-    if (loading) {
-      loadingMessage = (
-        <p className="loading-message">
-          Fetching country data
-          <span>.</span>
-          <span>.</span>
-          <span>.</span>
-        </p>
-      )
-    }
-
     return (
-      <>
-        {selectedCountries.map((country) => (
-          <CountryData data={country} key={slugify(country.country)} removeCountry={this.removeCountry} />
-        ))}
-        {loadingMessage}
-        <div>
-          <ErrorList errors={errors.__all__ || []} className="m-v-s" />
-        </div>
+      <div className="country-chooser-section" id="country-chooser-section">
         <div className={`country-chooser ${open ? 'open' : ''}`}>
           <span className="button--plus" />
           <button
@@ -174,49 +99,17 @@ class CountryChooser extends React.Component {
           </button>
           {countryInput}
         </div>
-      </>
+      </div>
     )
   }
 }
 
-function createCountryChooser({ element, ...params }) {
-  ReactDOM.render(<CountryChooser {...params} />, element)
-}
-
 CountryChooser.propTypes = {
-  selectedCountries: PropTypes.arrayOf(
-    PropTypes.shape({
-      export_duty: PropTypes.number,
-      country: PropTypes.string,
-      commodity_name: PropTypes.string,
-      utz_offset: PropTypes.string,
-      timezone: PropTypes.string,
-      last_year_data: PropTypes.shape({
-        year: PropTypes.string,
-        trade_value: PropTypes.string,
-        country_name: PropTypes.string,
-        year_on_year_change: PropTypes.string,
-      }),
-      corruption_perceptions_index: PropTypes.shape({
-        rank: PropTypes.number,
-        country_code: PropTypes.string,
-        country_name: PropTypes.string,
-        cpi_score_2019: PropTypes.number,
-      }),
-      easeofdoingbusiness: PropTypes.shape({
-        total: PropTypes.number,
-        year_2019: PropTypes.number,
-        country_code: PropTypes.string,
-        country_name: PropTypes.string,
-      }),
-    }).isRequired
-  ).isRequired,
   countryList: PropTypes.arrayOf(
     PropTypes.shape({
       value: PropTypes.string.isRequired,
       label: PropTypes.string.isRequired,
     })
   ).isRequired,
+  addCountry: PropTypes.func.isRequired,
 }
-
-export { CountryChooser, createCountryChooser }