+++ tests/browser/test_export_plan_target_markets.py
@@ -208,6 +208,7 @@ def test_can_see_target_markets_data(
     mock_get_export_plan_market_data.return_value = {
         'timezone': 'Australia/Sydney',
         'CPI': 10,
+        'sectors': ['Automotive']
     }
     mock_get_comtrade_last_year_import_data.return_value = {
         'last_year_data_partner': {
@@ -241,6 +242,7 @@ def test_can_see_target_markets_data(
                 },
             }
         ],
+        'sectors': ['Automotive']
     }
     mock_get_exportplan.return_value = data